pbp: cast before shifting in the range-decoder init - signed overflow on ordinary data
    unsigned int code = (in[1] << 24) | (in[2] << 16) | (in[3] << 8) | in[4];

in[1] is an unsigned char. The usual arithmetic conversions promote it to
int, so shifting left by 24 overflows a signed int for any byte from 0x80
up, which is undefined behaviour.

It fires on ordinary images rather than only corrupt ones. `in` is
self->buff_compressed, filled directly from the file by cdstream_read, so
roughly half of all compressed sectors reach CDAccess_PBP_decompress with
the high bit set in that byte.

Behaviour-preserving: the corrected expression returns exactly what the
old one returned on a two's-complement machine. The same shape was
verified exhaustively over all 2^32 byte patterns for the identical bug
in libchdr's get_bigendian_uint32_t, which is fixed separately in
libretro-common.

Found by grepping the pattern across the tree after fuzzing turned up the
libchdr instance under UBSan.

Not fixed here, reported instead. Three lines above, `unsigned char head
= in[0]` is tested with `if (head < 0)`, which is always false, so the
uncompressed-data fallback that branch guards is dead code. It predates
the C conversion in 6b695478 and comes in from upstream. Correcting the
test would activate a memcpy path that has evidently never run, on a
length taken from the file, and that wants real PBP images to validate
against rather than a drive-by change alongside an unrelated fix.

+++ mednafen/cdrom/CDAccess_PBP.c
@@ -3806,7 +3806,13 @@ static int CDAccess_PBP_decompress(struct CDAccess_PBP *self, unsigned char *out
    unsigned char head = in[0];
 
    unsigned int range = 0xFFFFFFFF;
-   unsigned int code = (in[1] << 24) | (in[2] << 16) | (in[3] << 8) | in[4];
+   /* Cast before shifting: in[1] is an unsigned char, promoted to int, so
+    * in[1] << 24 overflows a signed int for any byte from 0x80 up. `in`
+    * is buff_compressed, read straight out of the PBP by cdstream_read,
+    * so roughly half of all compressed sectors reach here with the high
+    * bit set - this fires on ordinary images, not only corrupt ones. */
+   unsigned int code = ((unsigned int)in[1] << 24) | ((unsigned int)in[2] << 16) |
+                       ((unsigned int)in[3] <<  8) |  (unsigned int)in[4];
 
    if (head < 0) /* Check if we have a valid starting byte. */
    {
